refactor(auth-plugin): remove deprecated v1 role and permission controllers.
