+++ api-legacy-adapter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -16,3 +16,4 @@
 
 com.alibaba.nacos.legacy.adapter.autoconfigure.LegacyAdapterServerWebAutoConfiguration
 com.alibaba.nacos.legacy.adapter.autoconfigure.LegacyAdapterConsoleAutoConfiguration
+com.alibaba.nacos.legacy.adapter.autoconfigure.LegacyAdapterAuthAutoConfiguration
