+++ plugin-default-impl/nacos-default-plugin-all/pom.xml
@@ -27,11 +27,11 @@
     
     <artifactId>nacos-default-plugin-all</artifactId>
     <dependencies>
-        <dependency>
-            <groupId>com.alibaba.nacos</groupId>
-            <artifactId>default-auth-plugin</artifactId>
-            <version>${project.version}</version>
-        </dependency>
+<!--        <dependency>-->
+<!--            <groupId>com.alibaba.nacos</groupId>-->
+<!--            <artifactId>default-auth-plugin</artifactId>-->
+<!--            <version>${project.version}</version>-->
+<!--        </dependency>-->
         <dependency>
             <groupId>com.alibaba.nacos</groupId>
             <artifactId>default-control-plugin</artifactId>
